pyproject.toml
[tool.poetry]
name = "stuff-counter"
version = "2.0.0"
description = "A bot that counts stuff when you feel lazy to"
authors = ["toolifelesstocode <luvuyosilubane50@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12" # so that crescent does not error
hikari = {extras = ["server"], version = "^2.0.0.dev120"}
hikari-crescent = "^0.6.0"
python-dotenv = "^1.0.0"
emoji = "^2.4.0"
taskipy = "^1.11.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
black = "^23.3.0"


[tool.taskipy.tasks]
start = "python -m src"

[tool.ruff]
line-length = 100
select = ["ALL"]
ignore = ["ANN101"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F403"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
